split engine into files
